backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.16
botocore==1.39.16
certifi==2025.7.14
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
fonttools==4.59.0
fpdf2==2.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
openai==1.98.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
gunicorn==21.2.0
flask
python-dotenv
reportlab==4.4.3
